backend/requirements.txt
asgiref==3.8.1
Django==4.2.28
django-graphql-jwt==0.4.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
promise==2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.13.2
celery==5.3.6
redis==4.5.5
flower==1.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
gunicorn==20.1.0
djangorestframework==3.14.0
flake8==7.1.0
whitenoise==6.5.0
